+++ src/Libraries/SmartStore.Services/Payments/IPaymentService.cs
@@ -1,4 +1,5 @@
 using System.Collections.Generic;
+using SmartStore.Core.Domain.Customers;
 using SmartStore.Core.Domain.Orders;
 using SmartStore.Core.Domain.Payments;
 using SmartStore.Core.Plugins;
@@ -13,10 +14,10 @@ namespace SmartStore.Services.Payments
         /// <summary>
         /// Load active payment methods
         /// </summary>
-        /// <param name="filterByCustomerId">Filter payment methods by customer; null to load all records</param>
-		/// <param name="storeId">Load records allows only in specified store; pass 0 to load all records</param>
+		/// <param name="customer">Filter payment methods by customer and apply payment method restrictions; null to load all records</param>
+		/// <param name="storeId">Filter payment methods by store identifier; pass 0 to load all records</param>
         /// <returns>Payment methods</returns>
-		IEnumerable<Provider<IPaymentMethod>> LoadActivePaymentMethods(int? filterByCustomerId = null, int storeId = 0);
+		IEnumerable<Provider<IPaymentMethod>> LoadActivePaymentMethods(Customer customer = null, int storeId = 0);
 
 		/// <summary>
 		/// Determines whether a payment method is active\enabled for a shop
@@ -39,24 +40,35 @@ namespace SmartStore.Services.Payments
 
 
 		/// <summary>
-		/// Gets a payment method by system name
+		/// Gets all payment method extra data
+		/// </summary>
+		/// <returns>List of payment method objects</returns>
+		IList<PaymentMethod> GetAllPaymentMethods();
+
+		/// <summary>
+		/// Gets payment method extra data by system name
 		/// </summary>
 		/// <param name="systemName">Provider system name</param>
 		/// <returns>Payment method entity</returns>
 		PaymentMethod GetPaymentMethodBySystemName(string systemName);
 
 		/// <summary>
-		/// Insert a payment method
+		/// Insert payment method extra data
 		/// </summary>
 		/// <param name="paymentMethod">Payment method</param>
 		void InsertPaymentMethod(PaymentMethod paymentMethod);
 
 		/// <summary>
-		/// Updates a payment method
+		/// Updates payment method extra data
 		/// </summary>
 		/// <param name="paymentMethod">Payment method</param>
 		void UpdatePaymentMethod(PaymentMethod paymentMethod);
 
+		/// <summary>
+		/// Delete payment method extra data
+		/// </summary>
+		/// <param name="paymentMethod">Payment method</param>
+		void DeletePaymentMethod(PaymentMethod paymentMethod);
 
 
 		/// <summary>